[CHORE/#466] 피드 프로필 패키지 구조 변경
