+++ presentation/feedprofile/src/main/java/com/hilingual/presentation/feedprofile/component/bottomsheet/ReportBlockBottomSheet.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hilingual.presentation.feedprofile.profile.component
+package com.hilingual.presentation.feedprofile.component.bottomsheet
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
