+++ presentation/feedprofile/src/main/java/com/hilingual/presentation/feedprofile/component/card/FeedProfileInfo.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hilingual.presentation.feedprofile.profile.component
+package com.hilingual.presentation.feedprofile.component.card
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -42,6 +42,7 @@ import com.hilingual.core.common.extension.noRippleClickable
 import com.hilingual.core.designsystem.R
 import com.hilingual.core.designsystem.component.image.NetworkImage
 import com.hilingual.core.designsystem.theme.HilingualTheme
+import com.hilingual.presentation.feedprofile.component.button.FeedUserActionButton
 
 @Composable
 internal fun FeedProfileInfo(
