+++ presentation/feedprofile/src/main/java/com/hilingual/presentation/feedprofile/follow/component/tab/FollowScreen.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hilingual.presentation.feedprofile.follow
+package com.hilingual.presentation.feedprofile.follow.component.tab
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -33,8 +33,8 @@ import com.hilingual.core.designsystem.component.pulltorefresh.HilingualPullToRe
 import com.hilingual.core.designsystem.theme.HilingualTheme
 import com.hilingual.core.ui.component.item.feed.UserActionItem
 import com.hilingual.presentation.feedprofile.follow.model.FollowItemModel
-import com.hilingual.presentation.feedprofile.profile.component.FeedEmptyCard
-import com.hilingual.presentation.feedprofile.profile.component.FeedEmptyCardType
+import com.hilingual.presentation.feedprofile.component.card.FeedEmptyCard
+import com.hilingual.presentation.feedprofile.component.card.FeedEmptyCardType
 import kotlinx.collections.immutable.ImmutableList
 
 @OptIn(ExperimentalMaterial3Api::class)
